+++ visualization/app/codeCharta/state/effects/nodeContextMenu/nodeContextMenuCard/nodeContextMenuCard.component.ts
@@ -1,18 +1,17 @@
 import { Component, OnInit } from "@angular/core"
 import { rightClickedCodeMapNodeSelector } from "../rightClickedCodeMapNode.selector"
 import { Observable } from "rxjs"
-import { CodeMapNode, CcState } from "../../../../codeCharta.model"
+import { CcState, CodeMapNode } from "../../../../codeCharta.model"
 import { Store } from "@ngrx/store"
 import { MatCard } from "@angular/material/card"
 import { MarkFolderRowComponent } from "../markFolderRow/markFolderRow.component"
 import { MatDivider } from "@angular/material/divider"
-import { MatButton } from "@angular/material/button"
 import { FocusButtonsComponent } from "../focusButtons/focusButtons.component"
 import { FlattenButtonsComponent } from "../flattenButtons/flattenButtons.component"
 import { HighlightButtonsComponent } from "../highlightButtons/highlightButtons.component"
 import { ExcludeButtonComponent } from "../excludeButton/excludeButton.component"
 import { AsyncPipe } from "@angular/common"
-import { LastPartOfNodePathPipe } from "./lastPartOfNodePath.pipe"
+import { CopyPathButtonComponent } from "../copyButton/copyPathButton.component"
 
 @Component({
     selector: "cc-node-context-menu-card",
@@ -23,13 +22,12 @@ import { LastPartOfNodePathPipe } from "./lastPartOfNodePath.pipe"
         MatCard,
         MarkFolderRowComponent,
         MatDivider,
-        MatButton,
         FocusButtonsComponent,
         FlattenButtonsComponent,
         HighlightButtonsComponent,
         ExcludeButtonComponent,
-        AsyncPipe,
-        LastPartOfNodePathPipe
+        CopyPathButtonComponent,
+        AsyncPipe
     ]
 })
 export class NodeContextMenuCardComponent implements OnInit {